First try to fix all linter errors and added write permission to ci.yaml

+++ src/main/java/ua/edu/ucu/apps/flowerstore/flower/Flower.java
@@ -15,7 +15,7 @@ public class Flower extends FlowerSpec {
     public Flower(
         double price, 
         FlowerColor color, 
-        int sepalLength, 
+        double sepalLength, 
         FlowerType flowerType) {
         super(color, sepalLength, flowerType);
         this.price = price;
